refactor(gui): modular widgets with card images
